requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
colorama==0.4.6
cryptography==44.0.3
decorator==4.4.2
distro==1.9.0
Django==5.2.5
django-environ==0.12.0
djangorestframework==3.16.1
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
firebase-admin==6.9.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.6.0
fsspec==2025.5.1
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
instagram-private-api==1.6.0.0
instagrapi==1.19.6
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
jwt==1.3.1
lazy_loader==0.4
MarkupSafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mysql-connector-python==9.3.0
narwhals==1.37.1
networkx==3.5
ngrok==1.4.0
ninja==1.11.1.4
numpy==2.2.5
oauth2client==4.1.3
openai==0.28.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
prettytable==0.7.2
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyairtable==3.1.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.18.0
pydantic==1.10.9
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
pyreadline3==3.5.4
Pyrebase4==4.8.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jwt==4.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-futures==1.0.2
requests-toolbelt==0.9.1
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
sherlock-project==0.15.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
stem==1.8.2
streamlit==1.45.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tk==0.1.0
tkcalendar==1.6.1
tkhtmlview==0.3.1
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.52.4
twilio==9.6.2
types-requests==2.32.4.20250611
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
